Updated draft based on AC, AB, and GitHub comments
Added an explicit list of the problems these best practices attempt to
solve

Made more explicit that that the criteria are most important at the
FPWD stage, but may be applied earlier

Incorporated language from those opposing mandatory incubation at the
AC meeting and tried to  stress that its the maturity of a proposal and
not the way it got to maturity that is being assessed.

Made incubation “strongly recommended” while attempting to defuse
opposition from those uncomfortable with mandating that incubation
happen in CGs.

Stressed that while pre Rec Track processes don’t mandate consensus,
mature proposals should indeed have broad consensus across relevant
stakeholders

Added a Process Considerations section to capture questions that have
come up in discussions but are not in scope for this document